+++ clang-tools-extra/test/clang-tidy/checkers/Inputs/Headers/cuda/cuda_runtime.h
@@ -1,0 +1,3 @@
+#include "cuda.h"
+
+cudaError_t cudaDeviceReset();